feat: mvc 패턴 적용
