+++ .eslintrc.json
@@ -6,7 +6,8 @@
                 "endOfLine": "auto"
             }
         ],
-        "no-constant-condition": ["error", { "checkLoops": false }]
+        "no-constant-condition": ["error", { "checkLoops": false }],
+        "no-unused-vars": "error"
     },
     "env": {
         "es6": true,
