+++ src/utils/rl-interface.js
@@ -1,0 +1,10 @@
+import * as readline from 'node:readline/promises';
+
+export const rl = readline.createInterface({
+    input: process.stdin,
+    output: process.stdout,
+});
+
+rl.on('close', () => {
+    process.exit();
+});
